requirements.txt
appdirs==1.4.3
args==0.1.0
certifi==2019.11.28
chardet==3.0.4
clint==0.5.1
distlib==0.3.0
dj-database-url==0.5.0
Django==2.2.13
django-braces==1.14.0
django-cors-headers==3.2.1
django-heroku==0.3.1
django-jinja==2.2.1
django-jinja2==0.1
django-model-utils==4.0.0
django-star-ratings==0.9.1
django-templated-mail==1.1.1
django-tinymce==2.8.0
django-webpack-loader==0.7.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djoser==2.0.3
filelock==3.0.12
gunicorn==20.0.4
idna==2.9
Jinja2==2.11.1
linecache2==1.0.0
MarkupSafe==1.1.1
mysqlclient==1.4.6
Pillow==7.0.0
pkginfo==1.5.0.1
psycopg2==2.8.5
PyJWT==1.7.1
pytz==2019.2
requests==2.23.0
requests-toolbelt==0.9.1
six==1.14.0
sqlparse==0.3.0
swapper==1.1.2.post1
traceback2==1.4.0
twine==1.7.4
unittest2==1.1.0
urllib3==1.25.8
virtualenv==20.0.10
whitenoise==5.0.1
